make levm use ethrex tracer call and logs
