+++ crates/common/common.rs
@@ -4,3 +4,4 @@ pub mod serde_utils;
 pub mod types;
 pub use bytes::Bytes;
 pub mod base64;
+pub mod tracing;
